+++ src/main/java/lotto/domain/LottoRank.java
@@ -1,4 +1,4 @@
-package lotto.domain.error;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
